requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
python-dotenv==1.0.0
pymongo==4.6.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pyyaml==6.0.1
pytest==7.4.4